Change int to double in task 17

+++ task-17/Program.cs
@@ -1,17 +1,17 @@
 ﻿Console.Clear();
 Console.WriteLine("Введите координату X != 0: ");
-int x = Convert.ToInt32(Console.ReadLine());
+double x = Convert.ToDouble(Console.ReadLine());
 while (x == 0)
 {
 	Console.WriteLine("Ошибка!\nВведите координату X != 0: ");
-	x = Convert.ToInt32(Console.ReadLine());
+	x = Convert.ToDouble(Console.ReadLine());
 }
 Console.WriteLine("Введите координату Y != 0: ");
-int y = Convert.ToInt32(Console.ReadLine());
+double y = Convert.ToDouble(Console.ReadLine());
 while (y == 0)
 {
 	Console.WriteLine("Ошибка!\nВведите координату Y != 0: ");
-	y = Convert.ToInt32(Console.ReadLine());
+	y = Convert.ToDouble(Console.ReadLine());
 }
 if (x > 0)
 	if (y > 0)
